==> notebooks/00_data_prep.Rmd <==
---
title: "Data preparation and Objectives"
author: "Benzon Carlitos Salazar"
date: "1/8/2022"
output: html_document
---

```{r setup}
library(magrittr)

white_wine_dataset <- read.csv(here::here("data-raw", "winequality-white.csv"), 
                               header = TRUE, 
                               sep = ";")
```

# Data Preparation
```{r clean_and_save}
white_col_list <-
  colnames(white_wine_dataset[1:11]) %<>%
  gsub("[[:punct:]]", "_", .) %>%
  gsub("[[:space:]]", "_", .)

renamed_white_wine <-
  paste0(tolower(white_col_list))
  
white_wine_dataset %>%
  dplyr::rename_at(1:11, ~renamed_white_wine) %>%
  saveRDS(here::here("data", "white_wine_dataset.rds"))
```

# Objectives
We want to make a predictive model on the quality of **white wine** based on chemical properties of wine.

The following analytical approaches are taken:

* Multiple regression: the *Quality* response is assumed to be a continuous variable and is predicted by the independent 
predictors, all of which are continuous.
* Regression Tree
* Classification of wines based on the chemical properties: Unsupervised analysis
* Classification of wines based on the chemical properties: K-nearest neighbors classification

```{r clean_up, include = FALSE}
rm(list = ls(all.names = TRUE)) # clear all objects including hidden objects
invisible(gc()) # free up memory
```
