==> notebooks/01_eda_and_preprocessing.Rmd <==
---
title: "Exploratory Data Analysis and Data Preprocessing"
author: "Benzon Carlitos Salazar"
date: "1/13/2022"
output: html_document
---

```{r setup, message = FALSE, warning = FALSE, include = FALSE}
library(magrittr)

white_wine <- readRDS(here::here("data", "white_wine_dataset.rds"))
attach(white_wine)
```

# Exploratory Data Analysis and Data Preprocessing
All white wine data set variables are summarized and univariate analysis with plots are shown.

## Histograms
We start with histograms to show the distribution of the variable values.
```{r histograms, echo=FALSE}
graphics::par(mfrow = c(2,2), oma = c(1, 1, 0, 0) + 0.1, mar = c(3, 3, 1, 1) + 0.1)

# Quality
graphics::barplot((table(quality)), 
                  col=c("slateblue4", "slategray", "slategray1", "slategray2", "slategray3", "skyblue4"))
graphics::mtext("Quality", side = 1, outer = F, line = 2, cex = 0.8)

# Fixed Acidity
MASS::truehist(fixed_acidity, h = 0.5, col = "slategray3")
graphics::mtext("Fixed Acidity", side = 1, outer = F, line = 2, cex = 0.8)

# Volatile Acidity
MASS::truehist(volatile_acidity, h = 0.5, col = "slategray3")
graphics::mtext("Volatile Acidity", side = 1, outer = F, line = 2, cex = 0.8)

# Citric Acid
MASS::truehist(citric_acid, h = 0.5, col = "slategray3")
graphics::mtext("Citric Acid", side = 1, outer = F, line = 2, cex = 0.8)

# Residual Sugar
MASS::truehist(residual_sugar, h = 0.5, col = "slategray3")
graphics::mtext("Residual Sugar", side = 1, outer = F, line = 2, cex = 0.8)

# Chlorides
MASS::truehist(chlorides, h = 0.5, col = "slategray3")
graphics::mtext("Chlorides", side = 1, outer = F, line = 2, cex = 0.8)

# Free Sulfur Dioxide
MASS::truehist(free_sulfur_dioxide, h = 0.5, col = "slategray3")
graphics::mtext("Free Sulfur Dioxide", side = 1, outer = F, line = 2, cex = 0.8)

# Total Sulfur Dioxide
MASS::truehist(total_sulfur_dioxide, h = 0.5, col = "slategray3")
graphics::mtext("Total Sulfur Dioxide", side = 1, outer = F, line = 2, cex = 0.8)

# Density
MASS::truehist(density, h = 0.5, col = "slategray3")
graphics::mtext("Density", side = 1, outer = F, line = 2, cex = 0.8)

# pH
MASS::truehist(ph, h = 0.5, col = "slategray3")
graphics::mtext("pH", side = 1, outer = F, line = 2, cex = 0.8)

# Sulphates
MASS::truehist(sulphates, h = 0.5, col = "slategray3")
graphics::mtext("Sulphates", side = 1, outer = F, line = 2, cex = 0.8)

# Alcohol
MASS::truehist(alcohol, h = 0.5, col = "slategray3")
graphics::mtext("Alcohol", side = 1, outer = F, line = 2, cex = 0.8)
```

## Boxplots
We also utilize boxplots to get a different view of the spread.
```{r boxplots, echo=FALSE}
graphics::par(mfrow = c(1, 5), oma = c(1, 1, 0, 0) + 0.1,  mar = c(3, 3, 1, 1) + 0.1)

# Quality
graphics::boxplot((table(quality)), 
                  col=c("slateblue4", "slategray", "slategray1", "slategray2", "slategray3", "skyblue4"),
                  pch = 19)
graphics::mtext("Quality", side = 1, line = 2, cex = 0.8)

# Fixed Acidity
graphics::boxplot(fixed_acidity, col = "slategray3", pch = 19)
graphics::mtext("Fixed Acidity", side = 1, line = 2, cex = 0.8)

# Volatile Acidity
graphics::boxplot(volatile_acidity, col = "slategray3", pch = 19)
graphics::mtext("Volatile Acidity", side = 1, line = 2, cex = 0.8)

# Citric Acid
graphics::boxplot(citric_acid, col = "slategray3", pch = 19)
graphics::mtext("Citric Acid", side = 1, line = 2, cex = 0.8)

# Residual Sugar
graphics::boxplot(residual_sugar, col = "slategray3", pch = 19)
graphics::mtext("Residual Sugar", side = 1, line = 2, cex = 0.8)

# Chlorides
graphics::boxplot(chlorides, col = "slategray3", pch = 19)
graphics::mtext("Chlorides", side = 1, line = 2, cex = 0.8)

# Free Sulfur Dioxide
graphics::boxplot(free_sulfur_dioxide, col = "slategray3", pch = 19)
graphics::mtext("Free Sulfur Dioxide", side = 1, line = 2, cex = 0.8)

# Total Sulfur Dioxide
graphics::boxplot(total_sulfur_dioxide, col = "slategray3", pch = 19)
graphics::mtext("Total Sulfur Dioxide", side = 1, line = 2, cex = 0.8)

# Density
graphics::boxplot(density, col = "slategray3", pch = 19)
graphics::mtext("Density", side = 1, line = 2, cex = 0.8)

# pH
graphics::boxplot(ph, col = "slategray3", pch = 19)
graphics::mtext("pH", side = 1, line = 2, cex = 0.8)

# Sulphates
graphics::boxplot(sulphates, col = "slategray3", pch = 19)
graphics::mtext("Sulphates", side = 1, line = 2, cex = 0.8)

# Alcohol
graphics::boxplot(alcohol, col = "slategray3", pch = 19)
graphics::mtext("Alcohol", side = 1, line = 2, cex = 0.8)
```

## Observations
Main observation is that all variables contain outliers.

* Based on the histogram for `Quality`, the most concentrated values are in [5, 6, 7]. Only a small proportion in in 
categories [3, 4, 8], and none in [1, 2, 9].
* `Fixed Acidity`, `Volatile Acidity`, `Citric Acid` all have outliers. If those outliers are eliminated, the 
distribution of the variables may take a more symmetric form.
* `Residual Sugar` has a positively skewed distribution. This distribution will remain true regardless of whether the 
outliers are removed.
* Most of the outliers are on the larger side.
* `Alcohol` shows an irregularly shaped distribution, but these outliers are not as pronounced as the others.

# Summary Statistics and Correlations
For this section, we utilize the `psych` package. It is a general purpose toolbox for experimental psychology. However, 
we are more so interested in the package's functions for multivariate analyses and scale construction using various 
analysis. For more information, visit [CRAN - Package psych](https://cran.r-project.org/web/packages/psych/index.html).

The observations with `psych::describe()` correlates with summary statistics.
```{r summary_stats}
summary(white_wine)
psych::describe(white_wine)
```

* Range is much larger compared to IQR.
* Mean is usually greater than the median.

These observations solidifies the existence of outliers, thus before any analysis is performed, the outliers must be 
taken care of.

Next, we look at bivariate analysis, including all pairwise scatterplot and correlation coefficients. Since our 
observations indicate a non-normal distribution, we consider using the Pearson correlation and Spearman Rank 
correlation.

### Table: Pearson's Correlation
```{r pearson_correlation}
stats::cor(white_wine[, -12])
```

### Table: Spearman Rank Correlation
```{r spearman_rank}
stats::cor(white_wine[, -12], method = "spearman")
```

Pearson's correlation and Spearman Rank correlations are relatively close, thus we consider the former. High 
correlations (>= 40% in absolute values) exists in the following variables in the Pearson Correlation table:

* `Total Sulfur Dioxide`
* `Density`
* `pH`
* `Alcohol`

### Scatterplot of predictors
```{r scatterplor}
graphics::pairs(white_wine[, -12], gap = 0, pch = 19, cex = 0.4, col = "darkblue")
title(sub = "Scatterplot of chemical attributes", cex = 0.8)
```

# Data Preparation
Possibly the most important step in data preparation is to identify outliers. Since we have a multivariate data, we 
consider only those points which do not have any predictore variable value to be outside of limits constructed by 
boxplots. Therefore, we apply the following rule:

* A predictor value is considered to be an outlier only if it is greater than Q3 + 1.5IQR, where Q3 is the third 
quartile and we multiple the `Inner Quartile Range` by 1.5.

This is because the extreme outliers are all on the higher end of the values and the distributions are all positively 
skewed.

We then randomly divide our data into Training Data and Test Data of equal sizes (50% each).
```{r data_preparation}
limit_out <- rep(0, 11)

for (i in 1:11) {
  tile1 <- stats::quantile(white_wine[,i], 0.75)
  tile2 <- stats::IQR(white_wine[,i], 0.75)
  limit_out[i] <- tile1 + 1.5*tile2
}

WhiteWineIndex <- matrix(0, 4898, 11)
for (i in 1:4898)
  for (j in 1:11) {
    if (white_wine[i, j] > limit_out[j])
      WhiteWineIndex[i, j] <- 1
  }
WWInd <- apply(WhiteWineIndex, 1, sum)
WhiteWineTemp <- cbind(WWInd, white_wine)
Indexes <- rep(0, 208)
j <- 1
for (i in 1:4898) {
  if (WWInd[i] > 0) {
    Indexes[j] <- i
    j <- j + 1
  }
  else
    j <- j
}

WhiteWineLib <- white_wine[-Indexes, ]   # Inside of Q3+1.5IQR
indexes = sample(1:nrow(WhiteWineLib), size = 0.5 * nrow(WhiteWineLib))
white_wine_train <- WhiteWineLib[indexes, ]
white_wine_test <- WhiteWineLib[-indexes, ]

white_wine_train %>%
  saveRDS(here::here("data", "white_wine_train.rds"))
white_wine_test %>%
  saveRDS(here::here("data", "white_wine_test.rds"))
```

```{r clean_up, include = FALSE}
detach(white_wine)
rm(list = ls(all.names = TRUE)) # clear all objects including hidden objects
invisible(gc()) # free up memory
```
